+++ src/lib.rs
@@ -1,5 +1,7 @@
 pub mod parameters;
 pub mod parameters_robots;
+
+#[cfg(feature = "allow_filesystem")]
 pub mod parameters_from_file;
 
 pub mod utils;
@@ -8,11 +10,17 @@ pub mod kinematics_impl;
 
 pub mod constraints;
 
+#[cfg(feature = "allow_filesystem")]
 pub mod urdf;
+#[cfg(feature = "allow_filesystem")]
+mod parameter_error;
+#[cfg(feature = "allow_filesystem")]
+mod simplify_joint_name;
 
 #[cfg(test)]
+#[cfg(feature = "allow_filesystem")]
 mod tests;
-mod parameter_error;
+
